Remove `ProductsToPickHelper` and move all methods to `ProductsToPickRowsService`
https://github.com/metasfresh/metasfresh/issues/6949
